+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 * [Added] Properties of STPConnectAccountParams are now mutable.
 
 ### PaymentSheet
-* [Added] Support for Alipay, OXXO, and PayNow with PaymentIntents.
+* [Added] Support for Alipay, OXXO, PayNow, and PromptPay with PaymentIntents.
 * [Added] Support for Cash App Pay with SetupIntents and PaymentIntents with `setup_future_usage`.
 
 ## 23.15.0 2023-08-28